+++ src/datasources/alarms/types/ListAlarms.types.ts
@@ -26,7 +26,11 @@ export enum AlarmsProperties {
   occurredAt = 'occurredAt',
   properties = 'properties',
   resourceType = 'resourceType',
+   // The Source field is derived from properties.system, with a fallback to
+   // properties.minionId. If neither property is available, the source defaults to
+   // an empty string.
   source = 'source',
+  // The State column is determined using the values of clear and acknowledged.
   state = 'state',
   // TODO(AB-3356927): Add transition property support
   transitionOverflowCount = 'transitionOverflowCount',